% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_drop_matrix}
\alias{generate_drop_matrix}
\title{Random Drop-out Matrix Generation}
\usage{
generate_drop_matrix(reads, dims, ratio = 0.01)
}
\arguments{
\item{reads}{reads that decide the dropping probabilities}

\item{dims}{Dimension of the drop matrix}

\item{ratio}{proportional constant}
}
\value{
A 0-1 drop-out matrix
}
\description{
For simulating the false zero, use this function to generate random drop-out matrix
}
\examples{

generate_drop_matrix(meta$reads,dims = c(20,50))


}
