% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ratio_density}
\alias{plot_ratio_density}
\title{Ratio Density plot}
\usage{
plot_ratio_density(otu_dna, otu_rna, with_zero = TRUE)
}
\arguments{
\item{otu_dna}{DNA OTU table}

\item{otu_rna}{RNA OTU table}

\item{with_zero}{indicator whether we should included the double zero peak}
}
\value{
A list of Density plot and double zero proportion
}
\description{
Plot the log ratio of four peak group
}
\examples{
humann_2 <- read_microbiome_data("data.bactRPK.marginal.DRNA.ZOE2.rds",eliminate_taxa_threshold=0.1)
plot_ratio_density(humann_2[[1]],humann_2[[2]])

}
