% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_microbiome_data}
\alias{read_microbiome_data}
\title{Microbiome data input}
\usage{
read_microbiome_data(file, eliminate_taxa_threshold = 0)
}
\arguments{
\item{file}{the name of the file which the data are to be read from. Data should be in list format with three dimensions: otu, taxa and meta.}

\item{eliminate_taxa_threshold}{threshold for deciding unexpressed level to be elminated}
}
\value{
A list of microbiome data after preprocessing
}
\description{
Reads a file in list format and eliminate invalid taxa and samples (preprocessing)
}
\examples{


humann_2 <- read_microbiome_data("data.bactRPK.marginal.DRNA.ZOE2.rds",eliminate_taxa_threshold=0.1)



}
